More cleanup, added patch fragment.
The pach directory contains a small part of the tileset patch which has been modified
to load palettes embedded in tileset files. The rest of the patch must be downloaded
from elsewhere for now. Do not use this version of the patch unless your tilesets have
been saved with a version of Tileset Tool from this repo.